refactor: the debugger is no longer just for libmagic
